+++ assets/javascripts/kitten/components/expandable/expand-board.stories.js
@@ -10,10 +10,10 @@ import {
 } from '@storybook/addon-knobs/react'
 import { StyleRoot } from 'radium'
 import { ExpandBoard } from 'kitten/components/expandable/expand-board'
-import { List } from 'kitten/components/lists/list'
 import { Grid, GridCol } from 'kitten/components/grid/grid'
-import { Text } from 'kitten/components/typography/text'
+import { ExpandBoardWithButtonItemList } from 'kitten/components/expandable/expand-board.examples'
 import COLORS from 'kitten/constants/colors-config'
+import { pxToRem } from 'kitten/helpers/utils/typography'
 
 storiesOf('Expandable', module)
   .addDecorator(withKnobs)
@@ -23,115 +23,72 @@ storiesOf('Expandable', module)
       'The ExpandBoard is a container that can expand on click. It can accept any content.',
     )(() => (
       <StyleRoot>
-        <Grid>
-          <GridCol offset="1" col="10">
-            <ExpandBoard>
-              <ExpandBoard.Button expandChildren="Lancez votre projet">
-                KissKissBankBank & Co
-              </ExpandBoard.Button>
-              <ExpandBoard.Content>
-                <div style={styles.expandBoard.defaultContent}>
-                  KissKissBankBank et 1 483 037 KissBankers vous aident à
-                  réaliser vos projets créatifs, associatifs et
-                  entrepreneuriaux. Participez à la naissance de projets
-                  inspirants.
-                </div>
-              </ExpandBoard.Content>
-            </ExpandBoard>
-          </GridCol>
-        </Grid>
+        <div style={styles.container}>
+          <Grid>
+            <GridCol offset="1" col="10">
+              <ExpandBoard withAnimation={boolean('withAnimation', true)}>
+                <ExpandBoard.Button
+                  expandChildren={text(
+                    'Expanded button text',
+                    'Lancez votre projet',
+                  )}
+                >
+                  {text('Button text', 'KissKissBankBank & Co')}
+                </ExpandBoard.Button>
+                <ExpandBoard.Content>
+                  <div style={styles.expandBoard.defaultContent}>
+                    {text(
+                      'Board content',
+                      `KissKissBankBank et 1 483 037 KissBankers vous aident à
+                        réaliser vos projets créatifs, associatifs et
+                        entrepreneuriaux. Participez à la naissance de projets
+                        inspirants.`,
+                    )}
+                  </div>
+                </ExpandBoard.Content>
+              </ExpandBoard>
+            </GridCol>
+          </Grid>
+        </div>
       </StyleRoot>
     )),
   )
   .add(
-    'ExpandBoard with list of ButtonItem',
+    'ExpandBoard with a List of ButtonItem',
     withInfo(
       'This is an example of ExpandBoard composition with a list of buttons as content.',
     )(() => (
       <StyleRoot>
-        <Grid>
-          <GridCol offset="1" col="10">
-            <ExpandBoard>
-              <ExpandBoard.Button expandChildren="Précisez votre choix">
-                Je soutiens
-              </ExpandBoard.Button>
-              <ExpandBoard.Content>
-                <List>
-                  <List.ButtonItem>
-                    <div style={styles.buttonListItem.wrapper}>
-                      <Text
-                        tag="p"
-                        weight="regular"
-                        color="font-1"
-                        size="tiny"
-                        style={styles.buttonListItem.base}
-                      >
-                        Taille XS
-                      </Text>
-                      <Text tag="small" color="font1" size="micro">
-                        Disponibilité: 10/30
-                      </Text>
-                    </div>
-                  </List.ButtonItem>
-                  <List.ButtonItem disabled>
-                    <div style={styles.buttonListItem.wrapper}>
-                      <Text
-                        tag="p"
-                        weight="regular"
-                        color="font-1"
-                        size="tiny"
-                        style={styles.buttonListItem.base}
-                      >
-                        Taille M
-                      </Text>
-                      <Text tag="small" color="font1" size="micro">
-                        Disponibilité: 10/30
-                      </Text>
-                    </div>
-                  </List.ButtonItem>
-                  <List.ButtonItem>
-                    <div style={styles.buttonListItem.wrapper}>
-                      <Text
-                        tag="p"
-                        weight="regular"
-                        color="font-1"
-                        size="tiny"
-                        style={styles.buttonListItem.base}
-                      >
-                        Taille XXL
-                      </Text>
-                      <Text tag="small" color="font1" size="micro">
-                        Disponibilité: 10/30
-                      </Text>
-                    </div>
-                  </List.ButtonItem>
-                </List>
-              </ExpandBoard.Content>
-            </ExpandBoard>
-          </GridCol>
-        </Grid>
+        <div style={styles.container}>
+          <Grid>
+            <GridCol offset="1" col="10">
+              <ExpandBoardWithButtonItemList
+                expandedButtonText={text(
+                  'Expanded button text',
+                  'Précisez votre choix',
+                )}
+                buttonText={text('Button text', 'Je soutiens')}
+                withAnimation={boolean('With animation', true)}
+              />
+            </GridCol>
+          </Grid>
+        </div>
       </StyleRoot>
     )),
   )
 
 const styles = {
+  container: {
+    margin: pxToRem(30),
+    paddingBottom: pxToRem(30),
+    borderBottom: `2px solid ${COLORS.line2}`,
+  },
   expandBoard: {
     defaultContent: {
       padding: '1.5em',
       borderLeft: `${COLORS.line1} 0.125rem solid`,
       borderRight: `${COLORS.line1} 0.125rem solid`,
       borderBottom: `${COLORS.line1} 0.125rem solid`,
-    },
-  },
-  buttonListItem: {
-    base: {
-      margin: 0,
-      padding: 0,
-      lineHeight: '1.2rem',
-    },
-    wrapper: {
-      margin: '1rem 0',
-      padding: 0,
     },
   },
 }